+++ HISTORY.md
@@ -4,6 +4,7 @@
 - Support gzipped cellular barcode files.
 - Support 10x V2 barcoding scheme. Thanks to @tomasgomes for the fix.
 - Re-enable streaming for cellular barcode filtering.
+- Add umi_matrix option to fasttagcount. This outputs a non-umi-deduped matrix of counts, useful for QC.
 
 ## 0.8.0
 - Fix `fasttagcount` off-by-one issue.